Start using MQTT to broadcast sensor data. Requires qmqtt library
qmqtt library is in a rough state, doesn't build as a Qt module yet. But
it is all around the best solution I found. Using it with mosquitto
server.

Tried: qremoteobject headache inducing and buggy, never got it to work.
Tried: actor-framework works but very ugly syntax
Tried: MPI very nice but not really meant for this sort of application
Tried: libmosquitto great if I weren't using Qt. Seems to require use of
multiple inheritance to use callback functions, when combined with Qt
lead to function name conflicts.